backend/requirements.txt
flask
flask-cors
kubernetes
python-dotenv
psycopg2-binary